+++ iOS/Dust/Dust/View/DustStateView.swift
@@ -23,62 +23,14 @@ class DustStateView: UIView {
 
         return dateFormatter.string(from: dateTime)
     }
-}
 
-struct Grade {
-    enum Kind: Int {
-        case good = 1, normal, bad, veryBad
+    func makeGradientLayer(in view: UIView, startColor: CGColor) {
+        let gradientLayer = CAGradientLayer()
+        gradientLayer.frame = view.bounds
+        gradientLayer.startPoint = CGPoint(x: 0.5, y: 0)
+        gradientLayer.endPoint = CGPoint(x: 0.5, y: 1)
+        gradientLayer.colors = [ startColor, UIColor.white.cgColor ]
+        gradientLayer.shouldRasterize = true
+        view.layer.addSublayer(gradientLayer)
     }
-
-    let kind: Kind
-    let label: String
-    let symbol: String
-    let color: UIColor
-}
-
-struct GradeFactory {
-    static let config: Dictionary<Grade.Kind, (label: String, symbol: String, color: UIColor)> = [
-        .good: ("좋음", "😀", UIColor(hex: "#3D85DD")!),
-        .normal: ("보통", "🙂", UIColor(hex: "#23BA46")!),
-        .bad: ("나쁨", "😷", UIColor(hex: "#FF8900")!),
-        .veryBad: ("매우 나쁨", "😱", UIColor(hex: "#C10404")!)
-    ]
-
-    static func create(by rawValue: Int?) -> Grade? {
-        guard let value = rawValue, let kind = Grade.Kind(rawValue: value) else { return nil }
-
-        return self.create(by: kind)
-    }
-
-    static func create(by kind: Grade.Kind) -> Grade {
-        guard let config = self.config[kind] else {
-            preconditionFailure("잘못된 값입니다: \(kind)")
-        }
-
-        return Grade(kind: kind, label: config.label, symbol: config.symbol, color: config.color)
-    }
-}
-
-extension DateFormatter {
-    static let yyyyMMdd: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd HH:mm"
-        formatter.calendar = Calendar(identifier: .iso8601)
-        formatter.timeZone = TimeZone(secondsFromGMT: 0)
-        formatter.locale = Locale(identifier: "ko_kr")
-
-        return formatter
-    }()
-
-    static let relativeDate: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.calendar = Calendar(identifier: .iso8601)
-        formatter.timeZone = TimeZone(secondsFromGMT: 9 * 3600)
-        formatter.locale = Locale(identifier: "ko_kr")
-        formatter.doesRelativeDateFormatting = true
-        formatter.dateStyle = .short
-        formatter.timeStyle = .short
-
-        return formatter
-    }()
 }
