+++ iOS/Dust/Dust/ViewController/ForecastViewController.swift
@@ -5,47 +5,7 @@ class ForecastViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-
-        if let blueColor = UIColor(hex: "#3D85DD")?.cgColor {
-            self.addGradientLayer(in: self.gradientView, startColor: blueColor)
-        }
-    }
-
-    func addGradientLayer(in view: UIView, startColor: CGColor) {
-        let gradientLayer = CAGradientLayer()
-        gradientLayer.frame = view.bounds
-        gradientLayer.startPoint = CGPoint(x: 0.5, y: 0)
-        gradientLayer.endPoint = CGPoint(x: 0.5, y: 1)
-        gradientLayer.colors = [ startColor, UIColor.white.cgColor ]
-        gradientLayer.shouldRasterize = true
-        view.layer.addSublayer(gradientLayer)
     }
 
 }
 
-extension UIColor {
-    public convenience init?(hex: String) {
-        let r, g, b: CGFloat
-
-        if hex.hasPrefix("#") {
-            let start = hex.index(hex.startIndex, offsetBy: 1)
-            let hexColor = String(hex[start...]).lowercased()
-
-            if hexColor.count == 6 {
-                let scanner = Scanner(string: hexColor)
-                var hexNumber: UInt64 = 0
-
-                if scanner.scanHexInt64(&hexNumber) {
-                    r = CGFloat((hexNumber & 0xff0000) >> 16) / 255
-                    g = CGFloat((hexNumber & 0x00ff00) >> 8) / 255
-                    b = CGFloat((hexNumber & 0x0000ff) >> 0) / 255
-
-                    self.init(red: r, green: g, blue: b, alpha: 1)
-                    return
-                }
-            }
-        }
-
-        return nil
-    }
-}
